+++ practica/demo/src/main/java/com/example/demo/dao/PaisDao.java
@@ -2,7 +2,8 @@ package com.example.demo.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.example.demo.model.Pais;
 
 @Repository
-public interface PaisDao extends JpaRepository<Pais, Long> {}
+public interface PaisDao extends JpaRepository<Pais, Integer> {}
