+++ practica/demo/src/main/java/com/example/demo/dao/ProvinciaDao.java
@@ -2,7 +2,8 @@ package com.example.demo.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.example.demo.model.Provincia;
 
 @Repository
-public interface ProvinciaDao extends JpaRepository<Provincia, Long> {}
+public interface ProvinciaDao extends JpaRepository<Provincia, Integer> {}
